+++ src/com/company/Engine/Items/Items.java
@@ -1,17 +1,19 @@
 package com.company.Engine.Items;
 
 
+import com.company.Engine.Player.Player;
+
 import java.util.ArrayList;
 
 public class Items {
     private short id;
-    private short drop;
+    private float drop;
     private ArrayList<Short> params;
 
     private String name;
     private String type;
 
-    public Items(short ID, String NAME, String TYPE, ArrayList<Short>PARAMS, short DROP) {
+    public Items(short ID, String NAME, String TYPE, ArrayList<Short>PARAMS, float DROP) {
         this.id = ID;
         this.name = NAME;
         this.type = TYPE;
@@ -31,11 +33,11 @@ public class Items {
         this.id = id;
     }
 
-    public short getDrop() {
+    public float getDrop() {
         return drop;
     }
 
-    public void setDrop(short drop) {
+    public void setDrop(float drop) {
         this.drop = drop;
     }
 
@@ -63,4 +65,6 @@ public class Items {
     public void setParams(ArrayList<Short> params) {
         this.params = params;
     }
+
+
 }
